Refactoring: Download button.
