+++ src/index.html
@@ -74,7 +74,7 @@
             </div>
 
 			<div class="btn-group" role="group">
-				<a class="btn btn-default js-tool-btn2" href="#" data-type="download" title="download">
+				<a id="downloadButton" class="btn btn-default" href="#" title="download">
 					<i class="fa fa-download"></i>
 				</a>
 			</div>
